+++ docs/changelog.rst
@@ -13,6 +13,7 @@ latest changes in development for next release
 
 * regular expression detection of Social Security Numbers (`#17`_)
 
+* Added functionality to keep ``replace_with = "identifier"`` (`#21`_)
 
 1.0.3
 -----
@@ -76,3 +77,4 @@ latest changes in development for next release
 .. _#12: https://github.com/datascopeanalytics/scrubadub/issues/12
 .. _#13: https://github.com/datascopeanalytics/scrubadub/issues/13
 .. _#17: https://github.com/datascopeanalytics/scrubadub/issues/17
+.. _#21: https://github.com/datascopeanalytics/scrubadub/issues/21